Fixed 'stuck picking up' bug, removed some old/unnecessary code
Fixed a bug where losing/winning the game while holding an item would cause you to not be able to pickup items in future levels. This was done by running the code in PlayerController.cs that drops items when the player presses the 'throw' button. I also removed chunks of old code in ButtonManager.cs that was unused, necessary, or even outright problematic. This should make the game loop more secure as well as simply any future debugging processes involving that script.