Moving platform fixes & alterations
The prefabs of the platforms & the path markers have been edited to fix their off-center transform. The platform will now go exactly where the gizmo is, and markers spawn in the correct location. Platforms can now also rotate when they turn corners (this is controlled by a bool in the inspector). The rotation speed scales with the amount of time each platform is set to pause at each marker to give the appearance of the laser marker guiding it along its path.