Fixed bug where beating tutorial wouldn't unlock level 1
This bug was a result of checking the beaten level for a number in it's name, i.e. "Ship Level 1", but the tutorial level does not have a number in it's name. Since this is the only level that will be this way (unless naming convention is broken), I fixed the bug by defaulting the number value to 0, since the tutorial is index 0 on the level list. The logic to retrieve a different value will only run if it returns true on a check for whether the scene name contains the string: "Tutorial". I also optimized the system so that other scripts only need to call the constructor, rather than the constructor and then a separate function depending on the script.